pyproject.toml
[tool.poetry]
name = "dataset-processing"
version = "1.0.0"
description = ""
authors = [
    "Alexander Dallmann <dallmann@informatik.uni-wuerzburg.de>",
    "Johannes Kohlmann <johannes.kohlmann@stud-mail.uni-wuerzburg.de>"
]

[tool.poetry.dependencies]
python = "^3.10"
typer = "^0.4.2"
pandas = "^1.4.3"
numpy = "^1.23.0"
pyspark = "^3.3.0"
tqdm = "^4.64.0"
psycopg2-binary = "^2.9.3"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
